Refactored code and started adding code for bible book dropdown
