--- src/components/Reading.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-export default function Reading() {
-  return (
-    <div>
-      <h1>Reading Plan </h1>
-    </div>
-  );
-}
