+++ src/components/read/TopBar.js
@@ -1,0 +1,121 @@
+import React from "react";
+import AppBar from "@material-ui/core/AppBar";
+import Button from "@material-ui/core/Button";
+import Toolbar from "@material-ui/core/Toolbar";
+import InputBase from "@material-ui/core/InputBase";
+import { fade, makeStyles } from "@material-ui/core/styles";
+import SearchIcon from "@material-ui/icons/Search";
+import Typography from "@material-ui/core/Typography";
+import IconButton from "@material-ui/core/IconButton";
+import { Link } from "react-router-dom";
+import MenuIcon from "@material-ui/icons/Menu";
+
+const useStyles = makeStyles(theme => ({
+  root: {
+    top: 0,
+    display: "flex",
+    width: "100%"
+  },
+  appBar: {
+    background: "#3970a7",
+    padding: "0px 20px"
+  },
+  title: {
+    flexGrow: 1,
+    display: "none",
+    [theme.breakpoints.up("sm")]: {
+      display: "block"
+    },
+    "& a": {
+      color: "inherit",
+      textDecoration: "none"
+    }
+  },
+  search: {
+    position: "relative",
+    borderRadius: theme.shape.borderRadius,
+    backgroundColor: fade(theme.palette.common.white, 0.15),
+    "&:hover": {
+      backgroundColor: fade(theme.palette.common.white, 0.25)
+    },
+    marginLeft: 0,
+    marginRight: theme.spacing(15),
+    width: "100%",
+    [theme.breakpoints.up("sm")]: {
+      marginLeft: theme.spacing(1),
+      width: "auto"
+    }
+  },
+  searchIcon: {
+    width: theme.spacing(7),
+    height: "100%",
+    position: "absolute",
+    pointerEvents: "none",
+    display: "flex",
+    alignItems: "center",
+    justifyContent: "center"
+  },
+  inputRoot: {
+    color: "inherit"
+  },
+  inputInput: {
+    padding: theme.spacing(1, 1, 1, 7),
+    transition: theme.transitions.create("width"),
+    width: "100%",
+    [theme.breakpoints.up("sm")]: {
+      width: 200,
+      "&:focus": {
+        width: 240
+      }
+    }
+  },
+  button: {
+    margin: theme.spacing(0),
+    color: "white",
+    borderColor: "white",
+    backgroundColor: "#007bff"
+  }
+}));
+export default function TopBar() {
+  const classes = useStyles();
+  return (
+    <div className={classes.root}>
+      <AppBar className={classes.appBar} position="static">
+        <Toolbar>
+          <IconButton
+            edge="start"
+            className={classes.menuButton}
+            color="inherit"
+            aria-label="Open drawer"
+          >
+            <MenuIcon />
+          </IconButton>
+          <Typography className={classes.title} variant="h5" noWrap>
+            <Link to="/">Vachanonline</Link>
+          </Typography>
+          <div className={classes.search}>
+            <div className={classes.searchIcon}>
+              <SearchIcon />
+            </div>
+            <InputBase
+              placeholder="Search…"
+              classes={{
+                root: classes.inputRoot,
+                input: classes.inputInput
+              }}
+              inputProps={{ "aria-label": "Search" }}
+            />
+          </div>
+          <Button
+            size="small"
+            variant="contained"
+            color="primary"
+            className={classes.button}
+          >
+            Login
+          </Button>
+        </Toolbar>
+      </AppBar>
+    </div>
+  );
+}
